# doc-files/kypo-elasticsearch-service-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 2.0.2"
  title: "KYPO Eleasticsearch Service - API Reference"
host: "localhost:8085"
basePath: "/kypo-elasticsearch-service/api/v1"
...
schemes:
- "http"
- "https"
paths:
...
  /training-platform-commands/access-token/{accessToken}/users/{userId}/ranges:
    get:
      tags:
      - "Training commands"
      summary: "Get aggregated commands of training specified by access token and\
        \ user identifier."
      description: ""
      operationId: "findAllConsoleCommandsByAccessTokenAndUserIdAndTimestampRange"
      produces:
      - "application/json"
      parameters:
      - name: "accessToken"
        in: "path"
        description: "Training instance access token"
        required: true
        type: "string"
      - name: "userId"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
        format: "int64"
      - name: "from"
        in: "query"
        description: "Lower bound of the time range (timestamp in epoch_millis format)\
          \ of the the resulting console commands."
        required: true
        type: "integer"
        format: "int64"
      - name: "to"
        in: "query"
        description: "Upper bound of the time range (timestamp in epoch_millis format)\
          \ of the the resulting console commands."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All training commands specified by access token and user identifier\
            \ were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /training-platform-commands/access-tokens/{accessToken}:
    get:
      tags:
      - "Training commands"
      summary: "Get all training commands specified by an access token."
      description: ""
      operationId: "findAllConsoleCommandsByAccessToken"
      produces:
      - "application/json"
      parameters:
      - name: "accessToken"
        in: "path"
        description: "Training access token."
        required: true
        type: "string"
      responses:
        200:
          description: "All commands in particular training instance were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Training commands"
      summary: "Delete all commands by access token."
      description: ""
      operationId: "deleteConsoleCommandsByAccessToken"
      produces:
      - "application/json"
      parameters:
      - name: "accessToken"
        in: "path"
        description: "Training instance access token"
        required: true
        type: "string"
      responses:
        200:
          description: "All events in particular training specified by access token\
            \ were deleted."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /training-platform-commands/access-tokens/{accessToken}/users/{userId}:
    get:
      tags:
      - "Training commands"
      summary: "Get all training commands specified by access token and user identifier."
      description: ""
      operationId: "findAllConsoleCommandsByAccessTokenAndUserId"
      produces:
      - "application/json"
      parameters:
      - name: "accessToken"
        in: "path"
        description: "Training instance access token"
        required: true
        type: "string"
      - name: "userId"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All training commands specified by access token and user identifier\
            \ were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Training commands"
      summary: "Delete all commands by access token and user identifier."
      description: ""
      operationId: "deleteConsoleCommandsByAccessTokenAndUserId"
      produces:
      - "application/json"
      parameters:
      - name: "accessToken"
        in: "path"
        description: "Training instance access token"
        required: true
        type: "string"
      - name: "userId"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All commands in particular training specified by access token\
            \ and user identifier were deleted."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /training-platform-commands/pools/{poolId}:
    get:
      tags:
      - "Training commands"
      summary: "Get all training commands specified by an pool."
      description: ""
      operationId: "findAllConsoleCommandsByPoolId"
      produces:
      - "application/json"
      parameters:
      - name: "poolId"
        in: "path"
        description: "Training pool ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All training commands specified by pool ID were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Training commands"
      summary: "Delete all commands by pool."
      description: ""
      operationId: "deleteConsoleCommandsByPoolId"
      produces:
      - "application/json"
      parameters:
      - name: "poolId"
        in: "path"
        description: "Training pool ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All events in particular training sandbox by id were deleted."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /training-platform-commands/sandboxes/{sandboxId}:
    get:
      tags:
      - "Training commands"
      summary: "Get all training commands specified by sandbox ID."
      description: ""
      operationId: "findAllConsoleCommandsBySandboxId"
      produces:
      - "application/json"
      parameters:
      - name: "sandboxId"
        in: "path"
        description: "Training sandbox ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All training commands specified by sandbox ID were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Training commands"
      summary: "Delete all commands by sandbox."
      description: ""
      operationId: "deleteConsoleCommandsByPoolId"
      produces:
      - "application/json"
      parameters:
      - name: "sandboxId"
        in: "path"
        description: "Training sandbox ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All commands in particular training specified by sandbox ID\
            \ were deleted."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /training-platform-commands/sandboxes/{sandboxId}/ranges:
    get:
      tags:
      - "Training commands"
      summary: "Get aggregated commands of training specified by sandbox ID."
      description: ""
      operationId: "findAllConsoleCommandsBySandboxIdAndTimestampRange"
      produces:
      - "application/json"
      parameters:
      - name: "sandboxId"
        in: "path"
        description: "Training sandbox ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "from"
        in: "query"
        description: "Lower bound of the time range (timestamp in epoch_millis format)\
          \ of the the resulting console commands."
        required: true
        type: "integer"
        format: "int64"
      - name: "to"
        in: "query"
        description: "Upper bound of the time range (timestamp in epoch_millis format)\
          \ of the the resulting console commands."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All training commands specified by sandbox ID were found."
          schema:
            type: "object"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
definitions:
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo2-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
...